requirements.txt
amqp==2.1.4
billiard==3.5.0.2
configparser==3.5.0
diff-match-patch==20121119
Django==1.10.4
django-braces==1.10.0
django-ckeditor==5.1.1
django-crispy-forms==1.6.1
django-import-export==0.5.1
enum34==1.1.6
flake8==3.2.1
kombu==4.0.2
mccabe==0.5.3
pep8==1.7.0
Pillow==3.4.2
psycopg2==2.6.2
pycodestyle==2.2.0
pyflakes==1.3.0
python-http-client==2.2.1
pytz==2016.10
sendgrid==3.6.3
slugify==0.0.1
tablib==0.11.3
vine==1.1.3
